Rename `transformation/transformers` to `transformation/visitors`
