+++ src/transformation/builtins/math.ts
@@ -1,8 +1,8 @@
 import * as ts from "typescript";
 import * as lua from "../../LuaAST";
 import { TransformationContext } from "../context";
-import { PropertyCallExpression, transformArguments } from "../transformers/call";
 import { UnsupportedProperty } from "../utils/errors";
+import { PropertyCallExpression, transformArguments } from "../visitors/call";
 
 export function transformMathProperty(node: ts.PropertyAccessExpression): lua.Expression {
     const name = node.name.text;
