+++ src/transformation/visitors/access.ts
