+++ src/transformation/visitors/class/index.ts
