+++ src/transformation/visitors/class/members/accessors.ts
