+++ src/transformation/visitors/class/members/fields.ts
