+++ src/transformation/visitors/class/members/method.ts
