+++ src/transformation/visitors/class/new.ts
