+++ src/transformation/visitors/class/setup.ts
