+++ src/transformation/visitors/class/utils.ts
