+++ src/transformation/visitors/delete.ts
