+++ src/transformation/visitors/loops/for-of.ts
