+++ src/transformation/visitors/namespace.ts
