+++ src/transformation/visitors/switch.ts
