+++ src/transformation/visitors/typeof.ts
