+++ src/transformation/visitors/variable-declaration.ts
